+++ modules/vfs/voltron/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.6
+
+- fix build error
+
+## 0.0.5
+
+- update ffi to 2.0.0
+
 ## 0.0.4
 
 - fix first reload error